block crosschain games (#113)

+++ backend/src/main/kotlin/com/retypeme/project/chain/ChainItemConfig.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class ChainItemConfig @JsonCreator constructor(
-    @JsonProperty("id") val id: Long,
+    @JsonProperty("id") val id: Int,
     @JsonProperty("name") val name: String,
     @JsonProperty("contract") val contract: String
 )
